Fixed slow rendering, added pugixml, fixed transform bugs
Need to work on scene description format

+++ crest3d/core/gfx/FrameBuffer.cpp
@@ -14,7 +14,7 @@ void FrameBuffer::initialise(float SCREEN_WIDTH, float SCREEN_HEIGHT)
 	// FULL Forward pass
 	glGenTextures(1, &framebufferTexture);
 	glBindTexture(GL_TEXTURE_2D, framebufferTexture);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, screenWidth, screenHeight, 0, GL_RGBA, GL_FLOAT, NULL);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, screenWidth, screenHeight, 0, GL_RGBA, GL_FLOAT, NULL);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
